requirements.txt
SQLAlchemy~=2.0.43
psycopg2-binary~=2.9.10
python-jose[cryptography]~=3.5.0
passlib[bcrypt]~=1.7.4
pydantic-settings~=2.10.1
python-multipart~=0.0.20
pypdf~=6.0.0
python-docx~=1.2.0

fastapi~=0.116.2
pip~=23.2.1
cryptography~=46.0.1
SQLAlchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
SQLAlchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
python-multipart
pypdf
python-docx
spacy~=3.8.7
torch~=2.8.0
sentence-transformers~=5.1.1
torch
pinecone-client~=6.0.0
google-generativeai

python-dotenv~=1.1.1
pydantic~=2.11.9
Jinja2~=3.1.6
rsa~=4.9.1
pyasn1~=0.6.1
numpy~=2.3.3
blis~=1.3.0
cffi~=2.0.0
setuptools~=80.9.0
typing_extensions~=4.15.0
lxml~=6.0.1
ecdsa~=0.19.1
rich~=14.1.0
Pygments~=2.19.2
markdown-it-py~=4.0.0
requests~=2.32.5
colorama~=0.4.6
sniffio~=1.3.1
cymem~=2.0.11
six~=1.17.0
mdurl~=0.1.2
charset-normalizer~=3.4.3
thinc~=8.3.6
srsly~=2.5.1
wasabi~=1.1.3
weasel~=0.4.1
typer~=0.17.4
click~=8.2.1
tqdm~=4.67.1
catalogue~=2.0.10
packaging~=25.0
langcodes~=3.5.0
preshed~=3.0.10
murmurhash~=1.0.13
pydantic_core~=2.33.2
confection~=0.1.5
shellingham~=1.5.4
MarkupSafe~=3.0.2
cloudpathlib~=0.22.0
certifi~=2025.8.3
starlette~=0.48.0
anyio~=4.10.0
urllib3~=2.5.0
greenlet~=3.2.4
annotated-types~=0.7.0
idna~=3.10
pycparser~=2.23
smart_open~=7.3.1
wrapt~=1.17.3
bcrypt~=4.3.0
spacy-legacy~=3.0.12
marisa-trie~=1.3.1
language_data~=1.3.0
spacy-loggers~=1.0.5
reportlab~=4.4.4
protobuf~=5.29.5
pillow~=11.3.0
scipy~=1.16.2
PyYAML~=6.0.2
filelock~=3.19.1
alembic~=1.16.5
uvicorn[standard]
pydantic[email]